moved lazy_load commands into plugins dir
